fixed error Resize Observer loop completed with undelivered notification
the issue was caused by the use of righ-0. Its weird right-anything seems to cause this error. top-* didnt show problems and left also was kinda okay but something like left-[60vw] would still cause it somewhat, as i remember. it may also be related to the use of vh (viewport height) when styling components. Anyways to get rid of absolute positionings i simply used a horizontal flexbox which can get the job done and wont even have to worry about positioning the components in the middle